Adjusted the INTERVAL type and references.
git-svn-id: http://deveeldb.googlecode.com/svn/trunk@153 de0d096c-ad4d-11de-a70b-4539db072d2e

+++ deveeldb/Deveel.Data.Sql/FetchStatement.cs
@@ -62,7 +62,7 @@ namespace Deveel.Data.Sql {
 				ResolveExpression(offsetExpr);
 
 				// and finally the value of the offset
-				offset = offsetExpr.Evaluate(null, context);
+				offset = (int) offsetExpr.Evaluate(null, context);
 			}
 
 			// so we finally fetch from the cursor
